Refactor GameAdapter into its own file
